requirements.txt
gradio>=4.0.0
chromadb>=0.5.0
litellm>=1.40.0
python-dotenv>=1.0.0
